Fixed up config file generator try.py, several small fixes

+++ tools/try.py
@@ -13,11 +13,12 @@ from django.template.loader import render_to_string
 from pysk.app.models import *
 from pysk.vps.models import *
 
-from time import time
+from time import time, sleep
 from IPy import IP
-from pwd import getpwnam
-from grp import getgrnam
+from pwd import getpwnam, getpwuid
+from grp import getgrnam, getgrgid
 from glob import glob
+from stat import S_IMODE, ST_MODE
 
 import cPickle
 import hashlib
@@ -30,6 +31,10 @@ import time
 import difflib
 
 validDomain = re.compile("([a-zA-Z0-9-]+\.?)+")
+DEBUG = True
+POSTMASTER_ADDRESS = "philipp@igowo.de"
+ABUSE_ADDRESS = "philipp@igowo.de"
+HYPERVISOR_IP = "188.40.56.202"
 
 # Utility functions
 def symlink(target, source):
@@ -37,12 +42,18 @@ def symlink(target, source):
     os.symlink(target, source)
 
 def rmtree(path):
-    print "rmtree        %s" % (path,)
-    shutil.rmtree(path)
+    if os.path.exists(path):
+        print "rmtree        %s" % (path,)
+        shutil.rmtree(path)
+    elif DEBUG:
+        print "rmtree SKIP   %s" % (path,)
 
 def remove(path):
-    print "remove        %s" % (path,)
-    os.remove(path)
+    if os.path.exists(path):
+        print "remove        %s" % (path,)
+        os.remove(path)
+    elif DEBUG:
+        print "remove SKIP   %s" % (path,)
 
 def copyfile(fromfile, tofile):
     print "copyfile      %s  %s" % (fromfile, tofile)
@@ -57,16 +68,28 @@ def runprog(args):
     subprocess.call(args)
 
 def chown(path, user, group=None):
-    print "chown         %s:%s  %s" % (user, group if group else user, path)
-    uid = getpwnam(user).pw_uid
-    gid = getpwnam(user).pw_gid
+    pw = getpwnam(user) if type(user) != int else getpwuid(user)
+    uid = pw.pw_uid
+    gid = pw.pw_gid
     if group != None:
-        gid = getgrnam(group).gr_gid
-    os.chown(path, uid, gid)
+        gid = getgrnam(group).gr_gid if type(group) != int else getgrgid(group)
+    
+    statinfo = os.stat(path)
+    if statinfo.st_uid != uid or statinfo.st_gid != gid:
+        print "chown         %s:%s  %s" % (user, group if group else user, path)
+        os.chown(path, uid, gid)
 
-def mkdir(path, mode):
-    print "mkdir    %o  %s" % (mode, path)
-    os.mkdir(path, mode)
+def chmod(path, mode):
+    statinfo = os.stat(path)
+    if S_IMODE(statinfo[ST_MODE]) != mode:
+        print "chmod    %o  %s" % (mode, path)
+
+def mkdir(path, mode, user, group=None):
+    if not os.path.exists(path):
+        print "mkdir    %o  %s" % (mode, path)
+        os.makedirs(path, mode)
+    chmod(path, mode)
+    chown(path, user, group)
 
 def makefile(path, content, mode=0644, strip_emptylines=False):
     print "makefile %o  %s" % (mode, path,)
@@ -74,7 +97,7 @@ def makefile(path, content, mode=0644, strip_emptylines=False):
         content = "\n".join([l for l in content.split("\n") if l.strip()])
     with open(path, "w") as conf:
         conf.write(content)
-    os.chmod(path, mode)
+    chmod(path, mode)
 
 def diff(oldfile, newfile, print_on_diff=True, move_on_diff=False, run_on_diff=None):
     print "diff          %s  %s" % (oldfile, newfile)
@@ -97,6 +120,8 @@ def diff(oldfile, newfile, print_on_diff=True, move_on_diff=False, run_on_diff=N
     if move_on_diff and len(output) > 0:
         copyfile(oldfile, oldfile + ".old")
         rename(newfile, oldfile)
+    elif move_on_diff and len(output) == 0:
+        os.remove(newfile)
     
     if print_on_diff and len(output) > 0:
         print output
@@ -124,16 +149,12 @@ def sync(ist, soll, start_func=None, stop_func=None):
                 stop_func(i)
 
 ### Clean-up old stuff
-if os.path.exists("/etc/nginx/conf/sites-available"):
-    rmtree("/etc/nginx/conf/sites-available")
-if os.path.exists("/etc/nginx/conf/sites-enabled"):
-    rmtree("/etc/nginx/conf/sites-enabled")
+rmtree("/etc/nginx/conf/sites-available")
+rmtree("/etc/nginx/conf/sites-enabled")
 
 ### GLOBALS
-
 fqdn = socket.getfqdn()
 ip = socket.gethostbyname_ex(socket.gethostname())[2][0]
-hypervisor_ip = "188.40.56.202"
 
 ### STARTUP
 monit_list = [os.path.basename(x) for x in glob("/opt/pysk/serverconfig/etc/monit.d/*")]
@@ -143,15 +164,12 @@ monit_restart_list = []
 ### SERVER CONFIG
 runprog(["/opt/pysk/serverconfig/copy.sh"])
 
-makefile("/etc/rc.conf", render_to_string("etc/rc.conf", {"hostname": fqdn, "ip": ip, "hypervisor_ip": hypervisor_ip}))
+makefile("/etc/rc.conf", render_to_string("etc/rc.conf", {"hostname": fqdn, "ip": ip, "hypervisor_ip": HYPERVISOR_IP}))
 
 ### DOVECOT
 statinfo = os.stat("/home/vmail/")
 vmail_uid = statinfo.st_uid
 vmail_gid = statinfo.st_gid
-
-if os.path.exists("/etc/dovecot/passwd.new"):
-    remove("/etc/dovecot/passwd.new")
 
 output = []
 for m in Mailbox.objects.filter(active=True).order_by("mail", "domain__name"):
@@ -167,71 +185,61 @@ d = diff("/etc/dovecot/passwd", "/etc/dovecot/passwd.new", move_on_diff=True)
 ### POSTFIX
 # virtual_mailboxes
 makefile("/etc/postfix/virtual_mailboxes.new", "\n".join(["%s@%s\t%s/%s/Maildir/" % (m.mail, m.domain, m.domain, m.mail) for m in Mailbox.objects.filter(active=True)]) + "\n")
-diff("/etc/postfix/virtual_mailboxes", "/etc/postfix/virtual_mailboxes.new", move_on_diff=True)
-runprog(["/usr/sbin/postmap", "/etc/postfix/virtual_mailboxes"])
+diff("/etc/postfix/virtual_mailboxes", "/etc/postfix/virtual_mailboxes.new", move_on_diff=True, run_on_diff=lambda: runprog(["/usr/sbin/postmap", "/etc/postfix/virtual_mailboxes"]))
 
 # virtual_domains
 domains = sorted(set([m.domain.name for m in Mailbox.objects.filter(active=True)]) | set([f.domain.name for f in Forwarding.objects.filter(active=True)]))
 makefile("/etc/postfix/virtual_domains.new", "\n".join(["%s\tdummy" % (d,) for d in domains]) + "\n")
-diff("/etc/postfix/virtual_domains", "/etc/postfix/virtual_domains.new", move_on_diff=True)
-runprog(["/usr/sbin/postmap", "/etc/postfix/virtual_domains"])
+diff("/etc/postfix/virtual_domains", "/etc/postfix/virtual_domains.new", move_on_diff=True, run_on_diff=lambda: runprog(["/usr/sbin/postmap", "/etc/postfix/virtual_domains"]))
 
 # virtual_forwardings
-POSTMASTER_ADDRESS = "philipp@igowo.de"
-ABUSE_ADDRESS = "philipp@igowo.de"
-
 forwardings = ["%s@%s\t%s" % (f.source, f.domain, f.target) for f in Forwarding.objects.filter(active=True)]
 forwardings += ["postmaster@%s\t%s" % (domain, POSTMASTER_ADDRESS) for domain in domains]
 forwardings += ["abuse@%s\t%s" % (domain, ABUSE_ADDRESS) for domain in domains]
 makefile("/etc/postfix/virtual_forwardings.new", "\n".join(forwardings) + "\n")
-diff("/etc/postfix/virtual_forwardings", "/etc/postfix/virtual_forwardings.new", move_on_diff=True)
-runprog(["/usr/sbin/postmap", "/etc/postfix/virtual_forwardings"])
+diff("/etc/postfix/virtual_forwardings", "/etc/postfix/virtual_forwardings.new", move_on_diff=True, run_on_diff=lambda: runprog(["/usr/sbin/postmap", "/etc/postfix/virtual_forwardings"]))
 
 ### APACHE
-# Generate an Apache monit-conf for every user who uses Apache
+mkdir("/srv/http/default/htdocs/", 0755, "root")
+
+# Generate an Apache instance + monit-conf for every user who uses Apache
 customers = set([x.owner for x in VirtualHost.objects.filter(apache_enabled=True)])
 for customer in customers:
     username = customer.user.username
     apacheroot = "/etc/httpd-%s" % (username,)
+    logpath = "/var/log/httpd-%s" % (username,)
+    runpath = "/var/run/httpd-%s" % (username,)
+
     ipoffset = customer.kundennr - 10000
     assert(ipoffset >= 0)
-    logpath = "/var/log/httpd-%s/" % (username,)
-    runpath = "/var/run/httpd-%s/" % (username,)
     
-    uid = getpwnam(username).pw_uid
-    gid = getpwnam(username).pw_gid
-
-    if not os.path.exists(logpath):
-        mkdir(logpath, 0755)
-    if not os.path.exists(runpath):
-        mkdir(runpath, 0755)
-    if not os.path.exists(runpath + "/fastcgi"):
-        mkdir(runpath + "/fastcgi", 0700)
-    os.chown(runpath + "/fastcgi", uid, gid)
-    if not os.path.exists(runpath + "/davlock"):
-        mkdir(runpath + "/davlock", 0700)
-    os.chown(runpath + "/davlock", uid, gid)
+    mkdir(logpath, 0755, "root")                                                            # /var/log/httpd-philwo
+    mkdir(runpath, 0755, "root")                                                            # /var/run/httpd-philwo
+    mkdir(os.path.join(runpath, "fastcgi"), 0700, username)                                 # /var/run/httpd-philwo/fastcgi
+    mkdir(os.path.join(runpath, "davlock"), 0700, username)                                 # /var/run/httpd-philwo/davlock
     
     if os.path.exists(apacheroot):
     	rmtree(apacheroot)
-    mkdir(apacheroot, 0755)
-    symlink("/usr/lib/httpd/build", apacheroot + "/build")
-    symlink("/usr/lib/httpd/modules", apacheroot + "/modules")
-    symlink(logpath, apacheroot + "/logs")
-    symlink(runpath, apacheroot + "/run")
-    mkdir(apacheroot + "/conf/", 0755)
-    mkdir(apacheroot + "/conf/sites/", 0755)
-    symlink("/etc/httpd/conf/magic", apacheroot + "/conf/magic")
-    symlink("/etc/httpd/conf/mime.types", apacheroot + "/conf/mime.types")
+    mkdir(apacheroot, 0755, "root")
+    symlink("/usr/lib/httpd/build", os.path.join(apacheroot, "build"))                      # /etc/httpd-philwo/build           -> /usr/lib/httpd/build
+    symlink("/usr/lib/httpd/modules", os.path.join(apacheroot, "modules"))                  # /etc/httpd-philwo/modules         -> /usr/lib/httpd/modules
+    symlink(logpath, os.path.join(apacheroot, "logs"))                                      # /etc/httpd-philwo/logs            -> /var/log/httpd-philwo
+    symlink(runpath, os.path.join(apacheroot, "run"))                                       # /etc/httpd-philwo/run             -> /var/run/httpd-philwo
+    mkdir(os.path.join(apacheroot, "conf"), 0755, "root")                                   # /etc/httpd-philwo/conf
+    mkdir(os.path.join(apacheroot, "conf", "sites"), 0755, "root")                          # /etc/httpd-philwo/conf/sites
+    symlink("/etc/httpd/conf/magic", os.path.join(apacheroot, "conf", "magic"))             # /etc/httpd-philwo/conf/magic      -> /etc/httpd/conf/magic
+    symlink("/etc/httpd/conf/mime.types", os.path.join(apacheroot, "conf", "mime.types"))   # /etc/httpd-philwo/conf/mime.types -> /etc/httpd/conf/mime.types
     
-    makefile(apacheroot + "/conf/httpd.conf", render_to_string("etc/httpd/conf/httpd.conf", {"username": username, "ipoffset": ipoffset}))
-    makefile(apacheroot + "/conf/sites/000-default", render_to_string("etc/httpd/conf/sites/default", {"ipoffset": ipoffset}))
-    makefile("/etc/monit.d/httpd-%s" % (username,), render_to_string("etc/monit.d/httpd", {"username": username, "ipoffset": ipoffset}))
+    makefile(os.path.join(apacheroot, "conf", "httpd.conf"),                                # /etc/httpd-philwo/conf/httpd.conf
+                            render_to_string("etc/httpd/conf/httpd.conf",
+                            {"username": username, "ipoffset": ipoffset}))
+    makefile(os.path.join(apacheroot, "conf", "sites", "000-default"),                      # /etc/httpd-philwo/conf/sites/000-default
+                            render_to_string("etc/httpd/conf/sites/default",
+                            {"ipoffset": ipoffset}))
+    makefile("/etc/monit.d/httpd-%s" % (username,),                                         # /etc/monit.d/httpd-philwo
+                            render_to_string("etc/monit.d/httpd",
+                            {"username": username, "ipoffset": ipoffset}))
     monit_list += ["httpd-%s" % (username,)]
-
-    if not os.path.exists("/srv/http/default/htdocs/"):
-    	os.makedirs("/srv/http/default/htdocs/", 0755)
-    os.chown("/srv/http/default/htdocs/", 0, 0)
     
     # For every IP of this server, which should be served by an apache
     for vh in VirtualHost.objects.filter(active=True, owner=customer, apache_enabled=True):
@@ -255,7 +263,7 @@ for customer in customers:
                 "apache_config": apache_config,
                 "enable_php": vh.enable_php,
             }))
-        makefile(apacheroot + "/conf/sites/%s" % (vh.fqdn(),), "\n".join(output))
+        makefile(os.path.join(apacheroot, "conf", "sites", vh.fqdn()), "\n".join(output))
 
     makefile("/etc/logrotate.d/httpd-%s" % (username,), render_to_string("etc/logrotate.d/httpd", {"username": username}), 0755)
     makefile("/usr/sbin/apachectl-%s" % (username,), render_to_string("usr/bin/apachectl", {"username": username, "ipoffset": ipoffset}), 0755)
@@ -268,7 +276,7 @@ makefile("/etc/nginx/conf/aliases", render_to_string("etc/nginx/conf/aliases", {
 
 if os.path.exists("/etc/nginx/conf/sites"):
     rmtree("/etc/nginx/conf/sites")
-mkdir("/etc/nginx/conf/sites", 0755)
+mkdir("/etc/nginx/conf/sites", 0755, "root")
 
 for vh in VirtualHost.objects.filter(active=True):
     ipoffset = vh.owner.kundennr - 10000
@@ -285,23 +293,14 @@ for vh in VirtualHost.objects.filter(active=True):
         makefile("/etc/nginx/conf/sites/%s-%s" % (vh.fqdn(), port), config, strip_emptylines=True)
 
     # Fixup htdocs dir
-    htdocs_dir = "/home/%s/www/%s/htdocs/" % (username, vh.fqdn())
-    uid = getpwnam(vh.owner.user.username).pw_uid
-    gid = getgrnam("users").gr_gid
-    
-    for dir in [os.path.realpath(htdocs_dir), os.path.realpath(os.path.join(htdocs_dir, "../"))]:
-        if not os.path.exists(dir):
-            print "WARNING: htdocs dir does not exist: %s" % (dir)
-            mkdir(dir, 0755)
-
-        statinfo = os.stat(dir)
-        if statinfo.st_uid != uid or statinfo.st_gid != gid:
-            chown(dir, vh.owner.user.username, "users")
+    vh_dir = "/home/%s/www/%s" % (username, vh.fqdn())
+    mkdir(vh_dir, 0755, vh.owner.user.username, "users")
+    mkdir(os.path.join(vh_dir, "htdocs"), 0755, vh.owner.user.username, "users")
 
 runprog(["/etc/rc.d/nginx", "reload"])
 
 ### PHP
-# Generate a PHP monit-conf for every user who uses PHP
+# Generate a PHP config + monit-conf for every user who uses PHP
 for customer in set([x.owner for x in VirtualHost.objects.filter(enable_php=True)]):
     username = customer.user.username
     makefile("/etc/php/php-%s.sh" % (username,), render_to_string("etc/php/php.sh", {"username": username, "php_instances": 5}), 0755)
@@ -311,9 +310,7 @@ for customer in set([x.owner for x in VirtualHost.objects.filter(enable_php=True
         "virtualhosts": VirtualHost.objects.filter(owner=customer, enable_php=True),
         "extensions": PHPExtension.objects.all(),
     }))
-    if not os.path.exists("/var/log/php-%s" % (username,)):
-        mkdir("/var/log/php-%s" % (username,), 0755)
-    chown("/var/log/php-%s" % (username,), username)
+    mkdir("/var/log/php-%s" % (username,), 0755, username)
     makefile("/etc/logrotate.d/php-%s" % (username,), render_to_string("etc/logrotate.d/php", {"username": username}))
     makefile("/etc/monit.d/php-%s" % (username,), render_to_string("etc/monit.d/php", {"username": username}))
     monit_list += ["php-%s" % (username,)]
@@ -326,9 +323,7 @@ makefile("/etc/php/php-pysk.ini", render_to_string("etc/php/php.ini", {
     "virtualhosts": None,
     "extensions": PHPExtension.objects.all(),
 }))
-if not os.path.exists("/var/log/php-pysk"):
-    mkdir("/var/log/php-pysk", 0755)
-chown("/var/log/php-pysk", "pysk")
+mkdir("/var/log/php-pysk", 0755, "pysk")
 makefile("/etc/logrotate.d/php-pysk", render_to_string("etc/logrotate.d/php", {"username": "pysk"}))
 makefile("/etc/monit.d/php-pysk", render_to_string("etc/monit.d/php", {"username": "pysk"}))
 monit_list += ["php-pysk"]
@@ -339,17 +334,16 @@ sync(glob("/etc/logrotate.d/php-*"), ["/etc/logrotate.d/php-%s" % (c.user.userna
 sync(glob("/etc/php/php-*.sh"), ["/etc/php/php-%s.sh" % (c.user.username,) for c in customers] + ["/etc/php/php-pysk.sh"], start_func=None, stop_func=remove)
 sync(glob("/etc/php/php-*.ini"), ["/etc/php/php-%s.ini" % (c.user.username,) for c in customers] + ["/etc/php/php-pysk.ini"], start_func=None, stop_func=remove)
 
+# Remove interfering config files (possibly installed by PHP distribution)
 for x in glob("/etc/php/conf.d/*"): remove(x)
 
 ### TARTARUS
-if not os.path.exists("/var/spool/tartarus"):
-    mkdir("/var/spool/tartarus", 0755)
-if not os.path.exists("/etc/tartarus"):
-    mkdir("/etc/tartarus", 0755)
-makefile("/etc/tartarus/secret.key", render_to_string("etc/tartarus/secret.key"), 0700)
-makefile("/etc/tartarus/home.conf", render_to_string("etc/tartarus/home.conf"), 0700)
-makefile("/etc/tartarus/etc.conf", render_to_string("etc/tartarus/etc.conf"), 0700)
-makefile("/etc/cron.d/tartarus", render_to_string("etc/cron.d/tartarus"), 0644)
+#mkdir("/var/spool/tartarus", 0755)
+#mkdir("/etc/tartarus", 0755)
+#makefile("/etc/tartarus/secret.key", render_to_string("etc/tartarus/secret.key"), 0700)
+#makefile("/etc/tartarus/home.conf", render_to_string("etc/tartarus/home.conf"), 0700)
+#makefile("/etc/tartarus/etc.conf", render_to_string("etc/tartarus/etc.conf"), 0700)
+#makefile("/etc/cron.d/tartarus", render_to_string("etc/cron.d/tartarus"), 0644)
 
 ### MONIT
 def monit_kill(id):
@@ -360,4 +354,5 @@ def monit_start(id):
 sync([os.path.basename(x) for x in glob("/etc/monit.d/*")], monit_list, monit_start, monit_kill)
 
 runprog(["/usr/bin/monit", "reload"])
+sleep(3)
 for x in monit_restart_list: runprog(["/usr/bin/monit", "restart", x])
